% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{cat_score}
\alias{cat_score}
\title{Pathway activity scores (PASs)}
\usage{
cat_score(
  x,
  gene_sets,
  species = "human",
  method = "AUCell",
  ncores = 1,
  return_df = FALSE,
  ...
)
}
\arguments{
\item{x}{Gene expression data which can be given either as a Seurat object,
or as a matrix of expression values where rows correspond to genes and
columns correspond to samples. This matrix can be also in a sparse format,
as a dgCMatrix.}

\item{gene_sets}{A list of vectors of genes for biological pathway.}

\item{species}{Species name, such as human or mouse.}

\item{method}{Method to employ in estimation of pathway activity scores
(PASs). By default this is set to AUCell (Sara Aibar et al., Nature
Methods, 2017) and other options are VISION (David DeTomaso et al.,
Nature Communications, 2019), ssGSEA (David A. Barbie et al., Nature, 2009),
or GSVA (Sonja Hänzelmann et al., BMC Bioinformatics, 2013).}

\item{ncores}{Number of cores to use for computation.}

\item{return_df}{Whether to return a data.frame.}

\item{...}{Other arguments.}
}
\value{
A data.frame of pathway activity scores.
}
\description{
Pathway activity scores (PASs)
}
\examples{
\dontrun{
data("pbmc_small")
score_data <- pbmc_small |>
  cat_score(
    gene_sets = "CP:KEGG",
    method = "gsva"
  )
}
}
