% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{compare_pathways}
\alias{compare_pathways}
\title{Analysis of differential pathway activity scores}
\usage{
compare_pathways(x, treatment, control, group_by = NULL, metadata = NULL)
}
\arguments{
\item{metadata}{}
}
\description{
Analysis of differential pathway activity scores
}
\examples{
\dontrun{
data("pbmc_small")

pbmc_small <- pbmc_small |>
  cat_score(gene_sets = "CP:KEGG", method = "GSVA")
pbmc_small

diff_pathway <- compare_pathways(
  pbmc_small,
  control = "g2",
  treatment = "g1",
  group_by = "groups"
)
}
}
