% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{cat_score}
\alias{cat_score}
\title{Pathway activity scores (PASs)}
\usage{
cat_score(
  x,
  gene_sets,
  species = "human",
  method = "AUCell",
  ncores = 1,
  return_seurat_obj = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Pathway activity scores (PASs)
}
\examples{
\dontrun{
data("pbmc_small")
score_data <- pbmc_small |>
  cat_score(gene_sets = "CP:KEGG",
            method = "gsva")
}
}
