% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{compare_pathways}
\alias{compare_pathways}
\title{Analysis of differential pathway activity scores}
\usage{
compare_pathways(x, treatment, control, group_by = NULL, metadata = NULL)
}
\arguments{
\item{x}{Pathway activity scores which can be given either as a Seurat
object, or as a matrix of expression values where rows correspond to
pathways and columns correspond to samples.}

\item{treatment}{Treatment group name.}

\item{control}{Control group name}

\item{group_by}{Column names in the metadata that contains grouping
information for the treatment and control groups.}

\item{metadata}{Metadata for each sample. Should be a data.frame where the
rows are sample names and the columns are additional metadata fields. Row
names in the metadata need to match the column names of the
pathway activity scores.}
}
\value{
A data.frame of differential pathway activity scores.
}
\description{
Analysis of differential pathway activity scores
}
\examples{
\dontrun{
data("pbmc_small")

pbmc_small <- pbmc_small |>
  cat_score(gene_sets = "CP:KEGG", method = "GSVA")
pbmc_small

diff_pathway <- compare_pathways(
  pbmc_small,
  control = "g2",
  treatment = "g1",
  group_by = "groups"
)
}
}
